correct SQL query for filtering out dictionary compressable datasets

+++ src/util/basic_fsst.h
@@ -213,7 +213,7 @@ inline void RunBasicFSST(duckdb::Connection &con, StringCollection &input, const
 
     
     global.run_time_ms = std::chrono::duration<double, std::milli>(end_time - start_time).count();
-    global.compression_factor = static_cast<double>(total_string_size) / total_compressed_string_size;
+    global.compression_factor = static_cast<double>(total_string_size) / static_cast<double>(total_compressed_string_size);
     
     PrintCompressionStats(total_strings_amount, total_string_size, total_compressed_string_size);
